# apps/api-worker/openapi/worker-foundation.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "SessionCookie": {
        "description": "Server-managed authenticated session cookie.",
        "in": "cookie",
        "name": "vocanova_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Operational VocaNova API running on Cloudflare Workers and D1.",
    "title": "VocaNova Worker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/canonical-words/{wordSlug}": {
      "get": {
        "operationId": "GetCanonicalWord",
        "parameters": [
          {
            "in": "path",
            "name": "wordSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "word": {
                      "properties": {
                        "difficultyLevel": {
                          "type": "string"
                        },
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "meanings": {
                          "items": {
                            "properties": {
                              "examples": {
                                "items": {
                                  "properties": {
                                    "exampleText": {
                                      "type": "string"
                                    },
                                    "id": {
                                      "format": "uuid",
                                      "type": "string"
                                    },
                                    "situationLabel": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "exampleText"
                                  ],
                                  "type": "object"
                                },
                                "type": "array"
                              },
                              "id": {
                                "format": "uuid",
                                "type": "string"
                              },
                              "learnerDefinition": {
                                "type": "string"
                              },
                              "partOfSpeech": {
                                "type": "string"
                              },
                              "reviewState": {
                                "enum": [
                                  "due",
                                  "new",
                                  "learning",
                                  "reviewing",
                                  "mastered",
                                  "not_reviewing",
                                  null
                                ],
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "saved": {
                                "type": "boolean"
                              },
                              "shortDefinition": {
                                "type": "string"
                              },
                              "usageNotes": {
                                "items": {
                                  "properties": {
                                    "id": {
                                      "format": "uuid",
                                      "type": "string"
                                    },
                                    "noteText": {
                                      "type": "string"
                                    },
                                    "noteType": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "noteType",
                                    "noteText"
                                  ],
                                  "type": "object"
                                },
                                "type": "array"
                              },
                              "userWordId": {
                                "format": "uuid",
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "partOfSpeech",
                              "shortDefinition",
                              "saved",
                              "reviewState",
                              "examples",
                              "usageNotes"
                            ],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "text": {
                          "type": "string"
                        },
                        "wordType": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "text",
                        "slug",
                        "wordType",
                        "meanings"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "word"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Content and learning"
        ]
      }
    },
...
    "/api/v1/journey-situations": {
      "get": {
        "operationId": "ListJourneySituations",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "properties": {
                          "category": {
                            "type": "string"
                          },
                          "displayOrder": {
                            "type": "integer"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "levelBand": {
                            "type": "string"
                          },
                          "shortDescription": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "slug",
                          "title",
                          "shortDescription",
                          "category",
                          "displayOrder"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Content and learning"
        ]
      }
    },
    "/api/v1/journey-situations/{slug}": {
      "get": {
        "operationId": "GetJourneySituation",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "meanings": {
                      "items": {
                        "properties": {
                          "meaningId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "partOfSpeech": {
                            "type": "string"
                          },
                          "saved": {
                            "type": "boolean"
                          },
                          "shortDefinition": {
                            "type": "string"
                          },
                          "wordId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "wordSlug": {
                            "type": "string"
                          },
                          "wordText": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "meaningId",
                          "wordId",
                          "wordSlug",
                          "wordText",
                          "partOfSpeech",
                          "shortDefinition",
                          "saved"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "situation": {
                      "properties": {
                        "category": {
                          "type": "string"
                        },
                        "displayOrder": {
                          "type": "integer"
                        },
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "levelBand": {
                          "type": "string"
                        },
                        "shortDescription": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "slug",
                        "title",
                        "shortDescription",
                        "category",
                        "displayOrder"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "situation",
                    "meanings"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Content and learning"
        ]
      }
    },
...
    "/api/v1/reviews/due": {
      "get": {
        "operationId": "GetReviewsDue",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "properties": {
                          "meaningId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "partOfSpeech": {
                            "type": "string"
                          },
                          "reviewStep": {
                            "type": "integer"
                          },
                          "shortDefinition": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "userWordId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "wordId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "wordSlug": {
                            "type": "string"
                          },
                          "wordText": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "userWordId",
                          "meaningId",
                          "wordId",
                          "wordText",
                          "wordSlug",
                          "partOfSpeech",
                          "shortDefinition",
                          "status",
                          "reviewStep"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "type": "string"
                    },
                    "totalCount": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "totalCount"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Content and learning"
        ]
      }
    },
...
    "/api/v1/user-words/{meaningId}": {
      "delete": {
        "operationId": "UnsaveUserWord",
        "parameters": [
          {
            "in": "path",
            "name": "meaningId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Content and learning"
        ]
      }
    },
...
  },
...
}